In listview mode use the default implemention of moveCursor
KCategorizedView overrides the moveCursor to handle moving the cursor in
the grid; however it's own implementation doesn't handle key up/down
when it's not in an exact grid mode; which happens to also include list
mode.

In this case we don't need anything custom and the QListView
implementation works perfectly.

This fixes KPluginSelector keyboard navigation.