Deprecate KFilterProxySearchLine
Summary:
It is too primitive and simple for anything but the absolutely most basic things

It is fixed to FixedStrings, search in all columns and caseinsensitive. And it seems to
not really be fixeable. So just  mark it as deprecated and remove it at a later point.

Test Plan: It has approximately 6 users across the kde codebase, let's fix them one by one.

Reviewers: kossebau, dfaure

Reviewed By: dfaure

Subscribers: vkrause, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D14772