KWidgetItemDelegateEventListener: remove useless destroy event sending
Commit 0225d2895a6f260aa918370e66db483c7d85cf7f to kdelibs added the
QEvent::Destroy handling branch. There is no respective comment or
discussion seen on the event forwarding to the viewport, and at least
with today's info is not obvious what purpose it would have:
KWidgetItemDelegate itself does not expect any such event on the viewport,
and any external code would also not have any reason to see it suddenly
appear there?
Possibly was motivated by seeing the forwarding happening
for the other branches, where there is a motivation though.